chore(api-gateway): route 세분화
study service 라우팅 리팩토링 중
